requirements.txt
numpy==1.21.0
pandas==1.3.0
tensorflow==2.6.0
matplotlib==3.4.2
keras==2.6.0
scikit-learn==0.24.2
pyyaml==5.4.1
jupyter==1.0.0
requests
streamlit
python-dotenv
openai
yfinance
pandas
matplotlib
scikit-learn
tensorflow
numpy
keras
pytest
flask
base64
io
traceback
